update: Refactor

+++ app/source/mpu9250.h
@@ -172,7 +172,7 @@ typedef struct
   int (*i2c_write) (uint8_t slave_addr, uint8_t reg_addr, uint8_t *data, uint32_t len);
 
   // Delay
-  void (*delay) (uint16_t ms);
+  void (*delay) (uint32_t ms);
 }
 mpu9250_t;
